requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
amadeus==12.0.0
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.8.1
asttokens==2.4.1
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.21
botocore==1.38.21
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==1.0.10
clear==2.0.0
click==8.1.8
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
durationpy==0.10
eval_type_backport==0.2.2
executing==2.2.0
fasta2a==0.2.6
fastapi==0.115.9
fastavro==1.11.1
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-genai==1.16.1
googleapis-common-protos==1.70.0
gotrue==2.12.0
gradio==5.31.0
gradio_client==1.10.1
greenlet==3.2.2
griffe==1.7.3
groovy==0.1.2
groq==0.25.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
logfire==3.16.0
logfire-api==3.16.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.1
mdurl==0.1.2
mistralai==1.7.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
narwhals==1.40.0
nest-asyncio==1.6.0
npm==0.1.1
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.81.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-asyncpg==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-instrumentation-httpx==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
optional-django==0.1.0
orjson==3.10.18
ormsgpack==1.9.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
playwright==1.52.0
playwright-stealth==1.0.6
pluggy==1.6.0
postgrest==1.0.2
posthog==4.2.0
priority==2.0.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.2.6
pydantic-ai-examples==0.2.6
pydantic-ai-slim==0.2.6
pydantic-evals==0.2.6
pydantic-graph==0.2.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.11
s3transfer==0.12.0
safehttpx==0.1.6
selenium==4.33.0
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.45.3
storage3==0.11.3
streamlit==1.45.1
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.4
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
